feat(): remove bricks: inject-preview-agent and collect-used-contract
