+++ bricks/visual-builder/src/bootstrap.ts
@@ -3,7 +3,6 @@ import "./workbench-sidebar/index.js";
 import "./workbench-action/index.js";
 import "./workbench-action-list/index.js";
 import "./workbench-tree/index.js";
-import "./data-providers/inject-preview-agent.js";
 import "./workbench-history-action/index.js";
 import "./page-arch-node/index.js";
 import "./contract-select/index.js";
